created new component fot rating show-rating.jsx
