+++ client/components/more-details-my-books.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import ShowRating from './show-rating';
 
 export default class MoreDetailsMybooks extends React.Component {
   constructor(props) {
@@ -43,27 +44,4 @@ export default class MoreDetailsMybooks extends React.Component {
       </>
     );
   }
-}
-function ShowRating(rating) {
-  const stars = [];
-  for (let i = 1; i < 6; i++) {
-    if (rating - i >= 0) {
-      stars.push(
-        <i key={i} className="fa-solid fa-star"></i>
-      );
-    } else if (rating - i < 0 && rating - i > -1) {
-      stars.push(
-        <i key={i} className="fa-solid fa-star-half-stroke"></i>
-      );
-    } else {
-      stars.push(
-        <i key={i} className="fa-regular fa-star"></i>
-      );
-    }
-  }
-  return (
-    <>
-      {stars}
-    </>
-  );
 }
