+++ client/pages/more-details.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import AppContext from '../lib/app-context';
+import ShowRating from './show-rating';
 
 export default class MoreDetails extends React.Component {
   constructor(props) {
@@ -109,27 +110,4 @@ export default class MoreDetails extends React.Component {
   }
 }
 
-function ShowRating(rating) {
-  const stars = [];
-  for (let i = 1; i < 6; i++) {
-    if (rating - i >= 0) {
-      stars.push(
-        <i key={i} className="fa-solid fa-star"></i>
-      );
-    } else if (rating - i < 0 && rating - i > -1) {
-      stars.push(
-        <i key={i} className="fa-solid fa-star-half-stroke"></i>
-      );
-    } else {
-      stars.push(
-        <i key={i} className="fa-regular fa-star"></i>
-      );
-    }
-  }
-  return (
-    <>
-      {stars}
-    </>
-  );
-}
 MoreDetails.contextType = AppContext;
